+++ docs/source/releases.rst
@@ -15,6 +15,7 @@ Additions
 ~~~~~~~~~
 
 - Added support for custom guild emojis.
+- Added support for message reactions.
 - Added :attr:`Guild.me` property for retreiving bot member.
 
 Bug fixes